requirements.txt
python-telegram-bot[webhooks]==20.*
requests[socks]
python-dotenv
google-generativeai
Pillow